+++ TombEngine/Renderer/RendererDrawEffect.cpp
@@ -458,7 +458,9 @@ namespace TEN::Renderer
 							fire.position.z + spark->position.z * fire.size / 2),
 						GetInterpolationFactor());
 
-					auto orientation = TO_RAD(Lerp(spark->PrevRotAng << 4, spark->rotAng << 4, GetInterpolationFactor()));
+					short prevOrientation = spark->PrevRotAng << 4;
+					short currentOrientation = spark->rotAng << 4;
+					auto orientation = TO_RAD(prevOrientation + Geometry::GetShortestAngle(prevOrientation, currentOrientation) * GetInterpolationFactor());
 					auto scalar = Lerp(spark->PrevScalar, spark->scalar, GetInterpolationFactor());
 					auto size = Vector2::Lerp(
 						Vector2(fire.PrevSize * spark->PrevSize, fire.PrevSize * spark->PrevSize),