+++ CHANGELOG.md
@@ -30,6 +30,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 
 * Fixed incorrect dynamic range for vertex colors, ambient light, dynamic lights and particle effects.
 * Fixed flyby camera jitter by converting the spline type to floating-point.
+* Fixed occasional flame emitter sprite jitter.
 
 ### Lua API changes